# optimize schema analysis performance