Fix https://github.com/istio/istio/issues/48665#issuecomment-1880316526, optimize `mapassign` and `fmt.Sprintf` performance issues.

Benchmark Metadata result: 
```code
$ benchstat old.txt new.txt
goos: darwin
goarch: amd64
pkg: istio.io/istio/pkg/config/analysis/analyzers/schema
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
            │   old.txt   │              new.txt               │
            │   sec/op    │   sec/op     vs base               │
Metadata-12   348.0n ± 6%   146.3n ± 7%  -57.96% (p=0.002 n=6)

            │  old.txt   │              new.txt              │
            │    B/op    │    B/op     vs base               │
Metadata-12   208.0 ± 0%   176.0 ± 0%  -15.38% (p=0.002 n=6)

            │  old.txt   │              new.txt              │
            │ allocs/op  │ allocs/op   vs base               │
Metadata-12   5.000 ± 0%   3.000 ± 0%  -40.00% (p=0.002 n=6)
```