@istio-testing 2024-01-08
@wulianglongrd: new pull request created: #48686

<details>

In response to [this](https://github.com/istio/istio/pull/48683#issuecomment-1880514832):

>/cherrypick release-1.20


Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository.
</details>